Use constants for special item names
A constant better represents these values as having special meaning and makes it easier to find them all in our code. It also reduces possibility for errors due to mistype.

Left strings as is in tests because they should be matching these exact strings and not depend on implementation details.